Switched uart to native code and removed last vestiges of arduino codebase
