+++ src/serial.cpp
@@ -27,23 +27,25 @@
 #include "hexops.h"
 #include "timer.h"
 
-#ifdef ARDUINO
+//#ifdef ARDUINO
 // WORK IN PROGRESS - some of the routines will flag an unused parameter 'pab' warning.
-#include <SoftwareSerial.h>
-#include <Arduino.h>
-#else
+//#include <SoftwareSerial.h>
+//#include <Arduino.h>
+//#else
 #include "uart.h"
-#endif
+//#endif
 
 #include "serial.h"
+
+#ifdef INCLUDE_SERIAL
 
 static const char baud_attrib[] PROGMEM = "B=";
 // Global defines
 volatile uint8_t  ser_open = 0;
 
-#ifdef ARDUINO
-SoftwareSerial    serial_peripheral( 8, 9 );  // either can support interrupts, and are otherwise available.
-#endif
+//#ifdef ARDUINO
+//SoftwareSerial    serial_peripheral( 8, 9 );  // either can support interrupts, and are otherwise available.
+//#endif
 
 extern uint8_t buffer[BUFSIZE];
 
@@ -91,17 +93,16 @@ static uint8_t hex_ser_open(pab_t pab) {
               baud = atol( attrib + 2 );
             }
           }
-#ifdef ARDUINO
-          serial_peripheral.begin( baud );
-#else
+//#ifdef ARDUINO
+//          serial_peripheral.begin( baud );
+//#else
           uart_config(CALC_BPS(baud), UART_LENGTH_8, UART_PARITY_NONE, UART_STOP_1);
-#endif
+//#endif
           if ( att & OPENMODE_READ ) {
-#ifdef ARDUINO
-            serial_peripheral.listen(); // apply listener if we are expecting input
-#else
-            // TODO figure out non Arduino alternative
-#endif
+//#ifdef ARDUINO
+//            serial_peripheral.listen(); // apply listener if we are expecting input
+//#else
+//#endif
           }
           transmit_word( 4 );
           transmit_word( len );
@@ -154,11 +155,11 @@ static uint8_t hex_ser_read(pab_t pab) {
   if ( ser_open ) {
     // protect access via ser_open since serial_peripheral is not present
     // if ser_open = 0.
-#ifdef ARDUINO
-    bcount = serial_peripheral.available();
-#else
+//#ifdef ARDUINO
+//    bcount = serial_peripheral.available();
+//#else
     bcount = (uart0_data_available() ? 1 : 0); // TODO  need to implement true count
-#endif
+//#endif
     if ( bcount > pab.buflen ) {
       bcount = pab.buflen;
     }
@@ -179,11 +180,11 @@ static uint8_t hex_ser_read(pab_t pab) {
 
         bcount -= len;
         while ( len-- && rc == HEXSTAT_SUCCESS ) {
-#ifdef ARDUINO
-          rc = transmit_byte( serial_peripheral.read() );
-#else
+//#ifdef ARDUINO
+//          rc = transmit_byte( serial_peripheral.read() );
+//#else
           rc = transmit_byte( uart_getc() );
-#endif
+//#endif
         }
       }
       if ( rc != HEXERR_BAV ) {
@@ -218,11 +219,11 @@ static uint8_t hex_ser_write(pab_t pab) {
       if (rc == HEXSTAT_SUCCESS) {
         j  = 0;
         while (j < len) {
-#ifdef ARDUINO
-          serial_peripheral.write( buffer[ j++ ] );
-#else
+//#ifdef ARDUINO
+//          serial_peripheral.write( buffer[ j++ ] );
+//#else
           uart_putc(buffer[j++]);
-#endif
+//#endif
         }
       }
       len -= i;
@@ -313,9 +314,9 @@ void ser_register(registry_t *registry) {
 
 void ser_reset(void) {
   if ( ser_open ) {
-#ifdef ARDUINO
-    serial_peripheral.end();
-#endif
+//#ifdef ARDUINO
+//    serial_peripheral.end();
+//#endif
     ser_open = FALSE;
   }
   return;
@@ -325,3 +326,5 @@ void ser_reset(void) {
 void ser_init(void) {
   ser_open = FALSE;
 }
+
+#endif
